Column list selection for DataTables API
Add support for column list selection using DataTables server-side API.

DataTables API is passing a list of columns (column[i]) that may not match
with the table definition on the server. For each request we build a list
of columns as requested by the client, then use this list for building query.
Columns with data source set to null or undefined are ignored, but searching
and ordering column indexes are mapped to the column list in request.
Returned results are labeled to match the request.

Returning nested data sources is not supported.